Remove identity table
Now everything is in GxsIdGroup and there's a type field to search by certain type.

Also remove image type because Retroshare doesn't use it (it uses type 0) and it's easy enough to use it as a "black box" containing image data. Finding the type is trivial anyway and just used for decoding.