Improve identities transfers
Now identities are transferred properly when there's a change on one side. The receiver would actually send them instead of requesting the ones it needs.

Also removed the ugly item introspection. Now there's a transaction flag which is kept within the transaction so that it knows what items are in there.