Refactor the item creation system
Items are no longer listed in their respective service. Introspection is run on startup to automatically add them to a map to process incoming items.

There's no more need to specify the service for outgoing items (except for GxsExchange ones) as each item defines its own service and subtype.

Since GxsExchange items are shared, this is not possible for them so their case is handled separately.

The service registry system and item creation is now a proper bean.

This also fixes the group and message transfer subtype problem.