Refactor context menus
Now it's much easier to set up context menus for ListView and Treeview alike classes. It's also more efficient because only one context menu is added on the parent object and not its items. MenuItems can be selectively and dynamically disabled or the context menu can be disabled all together depending on the originating item.