Refactoring of the startup system
There's now a network service that handles the network up/down and the associated services. It also handles the current IP address, its changes and if there's no network address available.

The local port is persisted into the settings instead of the connection entries. Startup has been simplified as well.

Also because of the requirement of the local port being saved into the settings, the database is not compatible anymore and the profile, etc... need to be recreated (or migrated).