Fix connection closing upon errors
When an exception happens because of an I/O error, Xeres would think the connection was still there. It happened both on the Netty level but also on the persistence level. This prevented peering when one end didn't add the certificate yet but added it later and would make the peer appear as connected even though it wasn't (and it would prevent further connection).