Improve connections
When running in server mode (--server-only) do not try to connect to the peer when a certificate is added.

Fix hostname entries in the ShortInvite/Certificate being saved as I2P connection type. Because of that fix, the database format is again incompatible. Please export your profile, delete the database then import it again.