Add Spring Security
The REST endpoint is now secured by default. A secure password is automatically generated on startup and used by both the client and the server. If no password is set, the old behavior is preserved.

Secure the WebSocket messaging endpoint as well.

Add a way to send synchronous events. They're still asynchronous by default.

Add the --control-address argument to specify to which IP to bind to.

Add the --no-control-password argument to have an unprotected endpoint (useful for docker and the likes).

Add the --remote-password argument to specify a password when connecting remotely.

Add a secure password generator.

Add a version system to upgrade the preferences.

Add a UI to configure the remote password.