Improve context menus over downloads
Context menu entries are shown properly over downloads (not possible to remove twice, not possible to open/show a file being downloaded, etc...)

Don't show a resumed file as transferring when it's still searching.

Make files being downloaded as hidden and prefix them with a dot so that they don't get in the way.