Fix longstanding identity update
This very old bug is finally fixed. It assumed the equality of Gxs signatures are only based on their gxs id (wrong! a signature can change!) and since it was put in a set, an identity update (with a new signature) would never work. This now allows to change the identity picture, among other things.