Redesign messaging window
The friend's avatar is displayed on top.

Use a larger input string with wrapping.

Add support for multiple lines in the private chat. Use CTRL + enter to produce them.

Improve focusing. Private messages won't steal the focus when using another part of the app.

On Windows, the task bar will blink when receiving a message and the window is iconified.