Mitigate side effects of connection status when there are 2 locations
We are connected to location A and B which have the same profile. We connect to both, then connection B drops. What is the profile status? We keep it at Available as long as there's at least one location connected. It previously would mark it as offline.

Also remove the stack trace for disconnection, unless we're in 'dev' profile.