Improve chat initiation
The chat button tries to use the best location in that order: available, busy, away, disconnected.

Added a chat context menu which uses the same logic.

To select a location directly, use the chat context menu over the location list.

Added a context menu to attempt a connection over a location. It will select a random connection.