Improve image viewer
Images can be shown in full screen when clicking on them.

Instructions on how to dismiss the full screen image is shown (but only once).

Images are automatically scaled to the private message window size so that they're always fully visible.

The image scaling no longer takes the full screen but the best image size as devised by JavaFX depending on the DPI screen.

There's no longer a sound event when opening an own chat window.