Fix clipboard pasting
Because of the way JavaFX inspects the clipboard, sometimes CTRL-V would fail if the clipboard contains some complex data (but still a string or an image). Then it would work the second time it's pressed. We bypass JavaFX and process it directly from now on.

Also images can be pasted from the context menu.