Fix sounds not playing
It can happen when a relative path is set and the app is started from the auto start (CurrentDir is different, then). Add a workaround for it and fix the default settings.

Also remove the splash screen removal trick. While it worked fine, it prevented it from showing up when launching manually, and it was brittle anyway.

Also move the log file. It's now in the app subdir because of the currentdir mentioned above.