Improve window opening
When a window appears out of bounds, it's repositioned at the center of the primary screen. This gets rid of invisible windows.

Tested on Windows only. Not sure if it happens in the first place on other platforms. It's still a mistery to me why Windows allows to open windows offscreen at all... I guess Windows is just bad at windows handling (pun intended).

Closes #108