Improve image support
Uses TwelveMonkeys so that the following formats can be read:
- WebP
- BMP (.ico files)

Following are improved:
- JPEG

When loading a PNG file as avatar, it uses PNG internally so that the transparency is preserved.

Inline images are sent as either PNG or JPEG depending on if they have transparency or not. This saves bandwidth.