Add image selector class and default placeholder
The image selector class allows to simple select an image for a group, etc...

Add a PlaceHolderImageView that shows a place holder image when there's nothing to display.

Add fixes and workarounds for the dreaded Ikonli sizing problem.

Add the board creation window, more work on the board message creation.